src/pyproject.toml
[tool.poetry]
name = "streamlit-app"
version = "0.1.0"
description = ""
authors = ["pottava"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.17.0"
pyrebase4 = "^4.6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pip-audit = "^2.4.5"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
